requirements.txt
flask==2.3.2
flask-cors==6.0.0
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
python-dotenv==1.0.0
pyjwt==2.7.0
werkzeug==3.0.6
flask-talisman==1.0.0
psycopg2-binary==2.9.6
SQLAlchemy~=2.0.41
Flask-Limiter~=3.12
Flask-Login~=0.6.3